api_yamdb/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
importlib-metadata==1.7.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2022.7.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.12.0
gunicorn==20.0.4
psycopg2-binary==2.8.6